move render diffing logic into app component to prevent whole tree re renders Implement more fixes from PR feedback
