+++ .gitignore
@@ -5,3 +5,4 @@ coverage
 !/public/index.html
 !/public/fortmaticG6oWpqbFTZ3fc3g2.html
 /public-clay
+.DS_Store
