+++ packages/manager/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@3id/manager",
-  "version": "0.2.0",
+  "version": "0.1.6",
   "author": "Ceramic Studio",
   "license": "(Apache-2.0 OR MIT)",
   "homepage": "https://github.com/ceramicstudio/3id-connect#readme",
@@ -38,7 +38,7 @@
   },
   "dependencies": {
     "3id-did-provider": "^1.0.2",
-    "@3id/common": "^0.2.0",
+    "@3id/common": "^0.1.3",
     "@ceramicnetwork/blockchain-utils-linking": "^1.0.0",
     "@ceramicnetwork/http-client": "^1.0.0",
     "@ceramicstudio/idx": "^0.12.0",
@@ -53,7 +53,7 @@
     "uint8arrays": "^2.1.4"
   },
   "devDependencies": {
-    "@3id/test-utils": "^0.2.0",
+    "@3id/test-utils": "^0.1.3",
     "@ceramicnetwork/cli": "^1.0.0",
     "@ceramicnetwork/common": "^1.0.0",
     "@ceramicnetwork/stream-caip10-link": "^1.0.0",
