+++ packages/test-utils/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@3id/test-utils",
-  "version": "0.2.0",
+  "version": "0.1.3",
   "author": "Ceramic Studio",
   "license": "(Apache-2.0 OR MIT)",
   "private": true,
